Add files via upload
This is the 1.02 version of the board which has a couple of minor changes. Mostly the "missing resistor" error has been fixed, two small resistors were changed to a more P12 compatible value (error in the BOM) and the breather hole was added.

This is a complete system, based on the Michelle 4.0 design - and while the head CAN be snapped off and used separately, it's not recommended.